Fixed database not ready issue
Docker-compose intentionally has weak dependency startup ordering; they
believe that apps should be resilient to their components not being
available. Since I don't want to overhaul Placement to run without being
able to connect to the DB, I've added a script that will first try to
connect to the DB port, which will loop until the port is available,
meaning that the DB is ready.